tests: divided tests into seperate files, added new tests
